Updating license to match the old shortcut editor for nuke because I based my key handling off of his. Thanks @dbr Lots of moving things around, I should have made smaller commits.
The keys subpackage has now been moved into keyboard/keys.
In the keys package, I got colours working.

There is a new top level subpackage called resources which holds util
classes and some constants. This is where my KeySequenceObject is.

The KeySequenceObject is the class that I based off of Dan's old shortcut
editor. I have changed some things about it because I needed a slightly
different spec. I have included a linkback to Dan's git file.

KeyboardWidget got modified quite heavily.
It now has the idea of an "active_sequence" which is the key sequence
currently being editted. Eventually every key that is pressed will be part
of at least on keyset. This will allow me to store the currently set
hotkeys in a keysequence and turn them on or off at ease. I will also
capture clicks with this and put them in the active_sequence.
There is some example code in colour_keyboard just so I can test the
context_check_boxes working. There is also some new helper functions.
The function str_to_key_int takes a str key name and gives you the Qt.Key int for it.
The function get_key will retrieve the weakref key object from the key_map
based on a str name passed into it.
The function _key_str_list_from_sequence takes a QKeySequence and will
return a list of the str names of each key in it.
The function clear_active_sequence will unclick each key in the
active_sequence.
The function set_key takes a str key_name and an optional state. If the
state passed is not a boolean, it will default to a toggle beheviour.
The function toggle_key calls set_key and makes it toggle, it then resets
the active_sequence
The function set_sequence takes a QKeySequence and will clear the active
sequence and replace it with the passed one, if it is different. If it is
the same as active, we are assume you are toggling and we toggle every key
in it.

KeyBoard has had a few keys names change as well as removing a setChecked
call on keys that was doubling up on signals for me.

All of the context check boxes have been moved to keyboard/checkboxes.
All of the checkboxes now use the constants in resources to store their
colour that they emit.

Dialog has gotten a big update too.
It is now using our KeySequenceObject instead of it's old keyEvent
handling which was lengthy.